ConsultIQ_DeepLearning_GenAI/consultiq/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
librosa==0.10.1
numpy==1.24.4
soundfile==0.12.1
scipy==1.10.1
transformers==4.30.2
torchaudio==2.1.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.1.16
pydantic==1.10.13
httpx==0.27.0
huggingface_hub==0.14.1
tokenizers==0.13.3
pymongo
google-generativeai
